move automl/config to zouwu/config
