+++ pyzoo/test/zoo/automl/searcher/test_ray_tune_search_engine.py
@@ -20,7 +20,7 @@ from zoo.zouwu.model.forecast.model import ModelBuilder
 from zoo.zouwu.model.forecast.model import model_creator as LSTM_model_creator
 import torch
 import torch.nn as nn
-from zoo.automl.config.recipe import Recipe
+from zoo.zouwu.config.recipe import Recipe
 from ray import tune
 import pandas as pd
 import numpy as np
